Restructure td_inline.h to mirror qd_inline.h and move td inline operators there
Rework the triple-double inline layer so `td_inline.h` plays the same role
for `td_real` that `qd_inline.h` already plays for `qd_real`.

Before this change, `td_inline.h` existed only as a thin bridge for a few
mixed-mode operators, while a large part of the basic `td_real` operator
surface still lived in `src/td_real.cpp`. That made the td inline structure
inconsistent with the established QD layout used by `dd_real` and `qd_real`.

This change makes the td layer more parallel to the existing qd design:

- expand `include/qd/td_inline.h` from a small bridge header into the real
  inline implementation home for triple-double primitive operators
- move td inline-friendly implementations into `td_inline.h`, including:
  - normalization helpers
  - three-sum / quick accumulation helpers
  - td<->qd conversion helpers used by the inline layer
  - constructors and assignments suitable for inlining
  - core `td_real` arithmetic operators
  - self operators
  - mixed-mode operators with `dd_real` and `qd_real`
  - comparisons and basic conversion helpers
- keep `src/td_real.cpp` focused on non-inlined functionality such as:
  - parsing / formatting
  - transcendental functions
  - higher-level helpers that are not intended to be header inlines
- preserve the public API added earlier for mixed `dd_real` / `td_real` /
  `qd_real` usage
- keep behavior unchanged while making the code organization match the
  repository’s existing pattern more closely

Why:
- make `td_real` structurally consistent with `dd_real` and `qd_real`
- reduce the mismatch between declaration layout and implementation layout
- make the triple-double operator surface easier to review and maintain
- keep mixed-mode operator support available through the same inline-first
  style used by the rest of QD

Verification:
- `make -j$(nproc)`
- `env LD_LIBRARY_PATH=/home/maho/code/QD/src/.libs ./tests/.libs/qd_test -td`
- `make check`

Results:
- td-specific tests passed
- full test suite passed:
  `qd_test`, `pslq_test`, `c_test`, `huge`, and `f_test`
